fp_portfolio/requirements.txt
aiohttp==3.6.2
flask-admin==1.5.7
Flask==1.1.1
flask-apispec==0.7.0
flask-mongoengine==0.9.5
Flask-Cors==3.0.8
marshmallow==3.8.0
gunicorn==19.9.0
pytest==6.1.1
pytest-asyncio==0.14.0
pytest-forked==1.3.0
pytest-xdist==2.1.0
requests==2.22.0
workalendar==14.0.0
xlrd==1.2.0